+++ package.json
@@ -26,6 +26,7 @@
     "babel": "^6.5.2",
     "babel-cli": "^6.6.5",
     "babel-preset-es2015": "^6.6.0",
+    "babelify": "^7.2.0",
     "browserify": "^13.0.0",
     "chai": "^3.5.0",
     "mocha": "^2.4.5",
